Download mkcert from web
Removed the mkcert from the repository. Them will be downloaded automatically from the sserve executable.
This allow to run sserve after a go get github.com/daquinoaldo/sserve installation.
Only the platform-specific mkcert executable is downloaded and it is placed in the system app data folder.

Also, the localhost.crt and localhost.key file are now installed in the app data folder, instead of the cwd.
This allows to keep clean the cwd directory and to reuse the certificates.